refactor: logger
- naming
- embed into struct
